+++ far2l/src/plug/plugapi.hpp
@@ -138,7 +138,7 @@ int WINAPI farColorDialog(INT_PTR PluginNumber, WORD* Color, int bAddTransparent
 
 DWORD WINAPI farGetCurrentDirectory(DWORD Size,wchar_t* Buffer);
 
-intptr_t WINAPI farCallFar(intptr_t CheckCode, FarMacroCall* Data);
+int WINAPI farCallFar(int CheckCode, FarMacroCall* Data);
 
 int WINAPI farGetFileEncoding(const wchar_t *FileName);
